Implement GeoLine::get_as_curve3d
It seems sort of right, but the length reported by Curve3D::get_baked_length differs from the actual LENGTH attribute. Not sure if this is because of some peculiarity with the data or a floating point inaccuracy.

Regarding floating point inaccuracy: A method to get a curve with an offset will be added.