Remove coordinate transformation code
This ended up almost working, but not quite... as it turns out, the reason is that we were using data which is projected differently, so it can't actually work with this method.

The code was now simplified as much as possible given the assumption that it gets Webmercator data. For non-webmercator data, the old function with gdalwarp could be reused.