Prevent crash in #26
An empty list is now returned instead of doing something which kills Godot. However, this solution is not final since we only output the error to the console at the moment. It's not trivial to let the GDNative layer know what went wrong, we might have to go with an Exception? Though Exceptions are unpopular in game programming...