Restructure GeoFeatures and GeoImage
GeoLine and GeoPoint now inherit from GeoFeature. This change introduces some casting in the GeoLine and GeoPoint, since they cast the general gdal_feature from the GeoFeature into a LineFeature/PointFeature. Duplicate functions such as get_attribute, as well as the destructor (which deletes the gdal feature), were now cleaned up with this.

In the process, these classes were put into separate files (geofeatures for the vector features, geoimage for that class) to make everything a bit clearer and easier to find.