Add geometry_type to Feature for cleaner typing
Previously we had some hacky type-casting, since there was no other way to know which specific feature type (such as LineFeature) a given Feature was.

Now, a geometry_type was added, which can be NONE, LINE, etc. That way, this variable can be checked, and the casting can happen based on that. This variable could also be exposed further outwards to Godot.