Re-integrate Raster pyramids and refine Raster structure
The Native* classes serve to improve the decoupling between the extractors and Geodot.

Pyramid support was integrated into this new structure, but may become deprecated or part of a bigger system in the future.

The restructure is now largely complete - the RasterDemo runs! Pyramid support has yet to be tested, however. Also, some refinement is necessary now that the structure is done - some functions may be unneeded, duplicated, badly named, etc.