Re-add some internal_constructor calls
These `::__internal_constructor` calls seem to be requried in some cases to ensure that the object is deleted properly. In some places, they cause errors.

This might be a bug with godot-cpp. The issue for it there is 215, which is closed, but it may have regressed somehow or it wasn't solved for all cases. It would be good to find out exactly when it is still required.