Overhaul of class instancing
As noted in that commit, 84358f2 shouldn't have been necessary as this was fixed in godot-cpp. And it's true - it is fixed, we just had some issues in our code.

A lot of Reference instancing was unclean and possibly the cause for crashes. This has now been completely overhauled with the proper way to do it, no more workarounds anywhere.