Improvement and cleanup in raster padding
This seems to resolve all the issues previously mentioned!

There are two minor things left to do: exposing the "padding" value (but likely with a default) and fixing the minor visual bug that occurs in some directions (marked with FIXME). But it is quite well usable in its current state.