Add functionality to get datasets for layers (and names)
This allows users to retrospectively get the dataset which a layer was taken from, e.g. for display in an UI.

Layer name functionality was also slightly improved and properly documented.

On the way, the get_layer_from_dataset function was moved from VectorExtractor to the NativeDataset - that was a remnant of the old structure.