Remove pyramid tile functionality
It was already noted in the comments for a while that this will be deprecated; it was quite a large amount of code for a very special case that hasn't been maintained and used in a long time.

It is preferable to pre-process this special case with GDAL, e.g. by georeferencing the tiles and building a VRT, or merging them into a GeoTIFF. This way, all our Raster handling is consistent.