refactor: inherit GeoDataset and Geo*Layer from RefCounted instead of Resource
We decided that inheriting from Resource is misleading, since the objects cannot actually be used entirely like a resource: the Resource.duplicate function and others were not properly implemented, and the caching based on resource_path was actually undesirable because we want to be able to load the same object multiple times for thread-safety.

The other changes replace instances of resource_name and resource_path.