fix: queries where the raster is padded on all sides
because the blocks were in an `if` `else if` construct, padding on both sides did not work properly; only when the query was outside of the available data on one side (left OR right, top OR bottom) did it work correctly.

Slightly restructured the code to remove the else if, works now.