fix: inconsistencies when saving/loading layers
Upsert seems to be more consistent than manually doing Create/Update; specifically, the first feature is not saved with the prior approach, while Upsert works for all.

Also fixed a missing step in `save_modified_layer`: features from the original layer which have been modified (but not deleted) in the cache were not saved.